+++ core/src/components/footer/footer.tsx
@@ -1,7 +1,7 @@
 import { Component, ComponentInterface, Element, Host, Prop, h } from '@stencil/core';
+import { findIonContent, getScrollElement, printIonContentErrorMsg } from '@utils/content';
 
 import { getIonMode } from '../../global/ionic-global';
-import { componentOnReady } from '../../utils/helpers';
 
 import { handleFooterFade } from './footer.utils';
 
@@ -56,17 +56,19 @@ export class Footer implements ComponentInterface {
 
     if (hasFade) {
       const pageEl = this.el.closest('ion-app,ion-page,.ion-page,page-inner');
-      const contentEl = (pageEl) ? pageEl.querySelector('ion-content') : null;
+      const contentEl = (pageEl) ? findIonContent(pageEl) : null;
+
+      if (!contentEl) {
+        printIonContentErrorMsg(this.el);
+        return;
+      }
 
       this.setupFadeFooter(contentEl);
     }
   }
 
-  private setupFadeFooter = async (contentEl: HTMLIonContentElement | null) => {
-    if (!contentEl) { console.error('ion-footer requires a content to collapse. Make sure there is an ion-content.'); return; }
-
-    await new Promise(resolve => componentOnReady(contentEl, resolve));
-    const scrollEl = this.scrollEl = await contentEl.getScrollElement();
+  private setupFadeFooter = async (contentEl: HTMLElement) => {
+    const scrollEl = this.scrollEl = await getScrollElement(contentEl);
 
     /**
      * Handle fading of toolbars on scroll
@@ -102,7 +104,7 @@ export class Footer implements ComponentInterface {
           [`footer-collapse-${collapse}`]: collapse !== undefined,
         }}
       >
-        { mode === 'ios' && translucent &&
+        {mode === 'ios' && translucent &&
           <div class="footer-background"></div>
         }
         <slot></slot>