+++ core/src/components/modal/modal.tsx
@@ -1,4 +1,5 @@
 import { Component, ComponentInterface, Element, Event, EventEmitter, Host, Method, Prop, State, Watch, h, writeTask } from '@stencil/core';
+import { printIonWarning } from '@utils/logging';
 
 import { config } from '../../global/config';
 import { getIonMode } from '../../global/ionic-global';
@@ -159,6 +160,7 @@ export class Modal implements ComponentInterface, OverlayInterface {
 
   /**
    * If `true`, the modal can be swiped to dismiss. Only applies in iOS mode.
+   * @deprecated - To prevent modals from dismissing, use canDismiss instead.
    */
   @Prop() swipeToClose = false;
 
@@ -199,6 +201,23 @@ export class Modal implements ComponentInterface, OverlayInterface {
   onTriggerChange() {
     this.configureTriggerInteraction();
   }
+
+  /**
+   * TODO (FW-937)
+   * This needs to default to true in the next
+   * major release. We default it to undefined
+   * so we can force the card modal to be swipeable
+   * when using canDismiss.
+   */
+
+  /**
+   * Determines whether or not a modal can dismiss
+   * when calling the `dismiss` method.
+   *
+   * If the value is `true` or the value's function returns `true`, the modal will close when trying to dismiss.
+   * If the value is `false` or the value's function returns `false`, the modal will not close when trying to dismiss.
+   */
+  @Prop() canDismiss?: undefined | boolean | (() => Promise<boolean>);
 
   /**
    * Emitted after the modal has presented.
@@ -270,7 +289,7 @@ export class Modal implements ComponentInterface, OverlayInterface {
   }
 
   componentWillLoad() {
-    const { breakpoints, initialBreakpoint } = this;
+    const { breakpoints, initialBreakpoint, swipeToClose } = this;
 
     /**
      * If user has custom ID set then we should
@@ -284,7 +303,11 @@ export class Modal implements ComponentInterface, OverlayInterface {
     }
 
     if (breakpoints !== undefined && initialBreakpoint !== undefined && !breakpoints.includes(initialBreakpoint)) {
-      console.warn('[Ionic Warning]: Your breakpoints array must include the initialBreakpoint value.');
+      printIonWarning('Your breakpoints array must include the initialBreakpoint value.')
+    }
+
+    if (swipeToClose) {
+      printIonWarning('swipeToClose has been deprecated in favor of canDismiss.\n\nIf you want a card modal to be swipeable, set canDismiss to `true`. In the next major release of Ionic, swipeToClose will be removed, and all card modals will be swipeable by default.');
     }
   }
 
@@ -358,6 +381,27 @@ export class Modal implements ComponentInterface, OverlayInterface {
   }
 
   /**
+   * Determines whether or not the
+   * modal is allowed to dismiss based
+   * on the state of the canDismiss prop.
+   */
+  private async checkCanDismiss() {
+    const { canDismiss } = this;
+
+    /**
+     * TODO (FW-937) - Remove the following check in
+     * the next major release of Ionic.
+     */
+    if (canDismiss === undefined) { return true; }
+
+    if (typeof canDismiss === 'function') {
+      return canDismiss();
+    }
+
+    return canDismiss;
+  }
+
+  /**
    * Present the modal overlay after it has been created.
    */
   @Method()
@@ -396,7 +440,17 @@ export class Modal implements ComponentInterface, OverlayInterface {
 
     if (this.isSheetModal) {
       this.initSheetGesture();
-    } else if (this.swipeToClose) {
+
+    /**
+     * TODO (FW-937) - In the next major release of Ionic, all card modals
+     * will be swipeable by default. canDismiss will be used to determine if the
+     * modal can be dismissed. This check should change to check the presence of
+     * presentingElement instead.
+     *
+     * If we did not do this check, then not using swipeToClose would mean you could
+     * not run canDismiss on swipe as there would be no swipe gesture created.
+     */
+    } else if (this.swipeToClose || (this.canDismiss !== undefined && this.presentingElement !== undefined)) {
       this.initSwipeToClose();
     }
 
@@ -528,6 +582,15 @@ export class Modal implements ComponentInterface, OverlayInterface {
       return false;
     }
 
+    /**
+     * If a canDismiss handler is responsible
+     * for calling the dismiss method, we should
+     * not run the canDismiss check again.
+     */
+    if (role !== 'handler' && !await this.checkCanDismiss()) {
+      return false;
+    }
+
     /* tslint:disable-next-line */
     if (typeof window !== 'undefined' && this.keyboardOpenCallback) {
       window.removeEventListener(KEYBOARD_DID_OPEN, this.keyboardOpenCallback);
@@ -605,20 +668,20 @@ export class Modal implements ComponentInterface, OverlayInterface {
       return;
     }
 
-    moveSheetToBreakpoint(
-      this.el,
-      this.backdropEl!,
-      this.animation!,
-      this.gesture!,
-      this.breakpoints ?? [],
-      1 - this.currentBreakpoint!,
-      breakpoint,
-      this.backdropBreakpoint,
-      () => {
-        this.sheetOnDismiss();
-      },
-      (newBreakpoint: number) => {
-        this.setNewBreakpoint(newBreakpoint);
+    moveSheetToBreakpoint({
+      baseEl: this.el,
+      backdropEl: this.backdropEl!,
+      animation: this.animation!,
+      gesture: this.gesture!,
+      breakpoints: this.breakpoints ?? [],
+      offset: 1 - this.currentBreakpoint!,
+      newBreakpoint: breakpoint,
+      backdropBreakpoint: this.backdropBreakpoint,
+      closest: breakpoint,
+      canDismissBlocksGesture: this.canDismiss !== undefined && this.canDismiss !== true && this.breakpoints![0] === 0,
+      currentBreakpoint: this.currentBreakpoint!,
+      onDismiss: () => this.sheetOnDismiss(),
+      onBreakpointChange: (newBreakpoint: number) => this.setNewBreakpoint(newBreakpoint)
     });
   }